README: fix block syntax according to the four-space rule
It seems that GitHub interprets the indentation here different than
other Markdown parsers, which always count multiples of four spaces as
indentation. GitHub also accepts three spaces as indentation for the
list item. To be compatible with other Markdown implementations, apply
the four-space rule.

The code block here is inside a list block, so it must be indented by
four spaces more than the list block.

Signed-off-by: Roland Hieber <rohieb@rohieb.name>